refactor(ai-chat): split resolve-thread-state from idempotency stage (plan U6)
thread-idempotency.ts: 482 -> 320 LOC. Lift the existing-thread block
(thread resolution, draft-session load + cleanup, pending-event action
revision context) to resolve-thread-state.ts (281 LOC). Pure mechanical
move; same DB calls, same error flow.

Stage LOC budget: all stages <=320, most <200. Handler.ts at 1491 still
over plan's <400 target; remainder needs U4c (model+tools loop) lift.
Snapshot suite 3/3 byte-identical.
